SPACE-MANAGEMENT/CreateUpdate-Space/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.10